編集削除権限をつけた

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Services\TweetService;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int) $request->route('tweetId');
+        if (!$tweetService->checkOwnTweet($request->user()->id, $request->id())) {
+            throw new AccessDeniedHttpException();
+        }
+
         $tweet = Tweet::whereId($tweetId)->firstOrFail();
         $tweet->delete();
         return redirect()
